backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
Flask-SQLAlchemy
Flask-Mail
Werkzeug
cryptography
google-generativeai
SQLAlchemy
Flask-Mail
pymysql
Flask-Migrate